exchange_system/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
asttokens==3.0.0
attrs==25.1.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.0
Django==5.1.5
django-appconf==1.0.6
django-cors-headers==4.6.0
django-staticfiles==1.2.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
drf-yasg==1.21.8
executing==2.2.0
icecream==2.1.4
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mysqlclient==2.2.7
packaging==24.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.22.3
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
